+++ docs/whats_new.rst
@@ -4,7 +4,7 @@
 v0.3.0 (unreleased)
 ===================
 
-* Can now query even the transform catalogs for the target `urlpath` (see future update in docs)
+* Can now query even the transform catalogs for the target `urlpath` (see `docs <https://model-catalogs.readthedocs.io/en/latest/demo.html#urlpath:-model-output-source>`_)
 * Can explicitly follow the transform catalog to the target catalog with `cat.follow_target()`.
 * `select_date_range()` now explicitly replaces the `urlpath` of the target when the function is run rather than later when `to_dask()` is run.